refactor: canvas_page.dart를 기능별로 여러 파일로 분리
- lib/models/ 폴더 생성하여 모델 클래스들 분리
  - canvas_color.dart: 캔버스 색상 enum
  - custom_scribble_notifier.dart: 커스텀 스크리블 알림 클래스

- lib/widgets/canvas/ 폴더 생성하여 캔버스 관련 위젯들 분리
  - color_button.dart: 색상 버튼 위젯
  - canvas_toolbar.dart: 툴바 관련 위젯들 (색상, 스트로크, 필압 등)
  - canvas_info.dart: 캔버스 정보 표시 위젯
  - canvas_background.dart: 배경 이미지 위젯
  - canvas_actions.dart: 캔버스 액션 유틸리티 함수들

- canvas_page.dart 간소화 및 모듈화
- 기능 유지하면서 코드 구조 개선
- 재사용성 및 유지보수성 향상
