+++ lib/widgets/canvas/canvas_background.dart
@@ -1,0 +1,62 @@
+import 'package:flutter/material.dart';
+
+class CanvasBackground extends StatelessWidget {
+  const CanvasBackground({
+    required this.width,
+    required this.height,
+    super.key,
+  });
+
+  final double width;
+  final double height;
+
+  @override
+  Widget build(BuildContext context) {
+    // 내부 로직 구성 필요 - 그냥 PDF-to-Image 사용할까
+    return _buildPlaceholder();
+  }
+
+  /// 플레이스홀더 위젯 (배경 이미지가 없을 때 표시)
+  Widget _buildPlaceholder() {
+    return Container(
+      width: width,
+      height: height,
+      decoration: BoxDecoration(
+        color: Colors.grey[50],
+        border: Border.all(
+          color: Colors.grey[300]!,
+          width: 2,
+          style: BorderStyle.solid,
+        ),
+      ),
+      child: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(
+              Icons.picture_as_pdf,
+              size: 64,
+              color: Colors.grey[400],
+            ),
+            const SizedBox(height: 16),
+            Text(
+              'PDF 이미지가 로드될 예정입니다',
+              style: TextStyle(
+                color: Colors.grey[600],
+                fontSize: 16,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              '크기: $width x $height px',
+              style: TextStyle(
+                color: Colors.grey[500],
+                fontSize: 12,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
