+++ lib/widgets/canvas/color_button.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+
+class ColorButton extends StatelessWidget {
+  const ColorButton({
+    required this.color,
+    required this.isActive,
+    required this.onPressed,
+    this.outlineColor,
+    this.child,
+    this.tooltip,
+    super.key,
+  });
+
+  final Color color;
+
+  final Color? outlineColor;
+
+  final bool isActive;
+
+  final VoidCallback onPressed;
+
+  final Icon? child;
+
+  final String? tooltip;
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedContainer(
+      duration: kThemeAnimationDuration,
+      decoration: ShapeDecoration(
+        shape: CircleBorder(
+          side: BorderSide(
+            color: switch (isActive) {
+              true => outlineColor ?? color,
+              false => Colors.transparent,
+            },
+            width: 2,
+          ),
+        ),
+      ),
+      child: IconButton(
+        style: FilledButton.styleFrom(
+          backgroundColor: color,
+          shape: const CircleBorder(),
+          side: isActive
+              ? const BorderSide(color: Colors.white, width: 2)
+              : const BorderSide(color: Colors.transparent),
+        ),
+        onPressed: onPressed,
+        icon: child ?? const SizedBox(),
+        tooltip: tooltip,
+      ),
+    );
+  }
+}
